Added Loss Functions and Training Verbosity. Also fixed issue with Sigmoid Calculation

+++ MyMachineLearningLibrary/SigmoidActivationFunction.cs
@@ -28,7 +28,7 @@ namespace MyMachineLearningLibrary
 			{
 				for (int j = 0; j < m.ColoumnLength; j++)
 				{
-					result[i, j] = (double)(1.0 / (1.0 + Math.Exp(-m[i, j])));
+					result[i, j] = 1.0 / (1.0 + Math.Exp(-m[i, j]));
 				}
 			}
 			return result;
